requirements.txt
flask
Flask-PyMongo
python-dotenv
gunicorn
flask-cors
bcrypt
PyJWT
pika==1.3.2
pymemcache==4.0.0
